Interim check in -> broken!
Cascading delete issues from SolutionData to ProjectFileData and
ProjectData.ProjectData - note: ProjectData has the same
ProjectFileData in it.  Need to refactor.  Also, need to refactor
LibraryReferenceData and cross project references as well,

Many to many relationships might debrittleize it, or maybe less
ownership.

+++ ProjectDependencies/DataAccess/ProjectFileData.cs
@@ -6,7 +6,7 @@
     [DebuggerDisplay("Name = {Name}")]
     public class ProjectFileData
     {
-        public long ProjectFileDataId { get; set; }
+        public int ProjectFileDataId { get; set; }
 
         public string Name { get; set; } = string.Empty;
         public string ProjectPath { get; set; } = string.Empty;
